feat: MLIBZ-2140  Collection kmd grows exponentially every time a sync pull is called

+++ android-lib/src/main/java/com/kinvey/android/cache/RealmCache.java
@@ -20,7 +20,6 @@ import com.google.api.client.json.GenericJson;
 import com.kinvey.java.KinveyException;
 import com.kinvey.java.Query;
 import com.kinvey.java.cache.ICache;
-import com.kinvey.java.model.AggregateEntity;
 import com.kinvey.java.model.AggregateType;
 import com.kinvey.java.model.Aggregation;
 import com.kinvey.java.query.AbstractQuery;
@@ -46,6 +45,11 @@ import io.realm.Sort;
  * Created by Prots on 1/26/16.
  */
 public class RealmCache<T extends GenericJson> implements ICache<T> {
+
+    private static final String ACL = "__acl";
+    private static final String KMD = "__kmd";
+    private static final String ID = "_id";
+
     private String mCollection;
     private RealmCacheManager mCacheManager;
     private Class<T> mCollectionItemClass;
@@ -171,10 +175,10 @@ public class RealmCache<T extends GenericJson> implements ICache<T> {
                     .beginGroup();
             Iterator<String> iterator = ids.iterator();
             if (iterator.hasNext()) {
-                query.equalTo("_id", iterator.next());
+                query.equalTo(ID, iterator.next());
                 while (iterator.hasNext()) {
                     String id = iterator.next();
-                    query.or().equalTo("_id", id);
+                    query.or().equalTo(ID, id);
                 }
             }
             query.endGroup();
@@ -200,7 +204,7 @@ public class RealmCache<T extends GenericJson> implements ICache<T> {
         try {
             mRealm.beginTransaction();
             DynamicRealmObject obj = mRealm.where(TableNameManager.getShortName(mCollection, mRealm))
-                    .equalTo("_id", id)
+                    .equalTo(ID, id)
                     .greaterThanOrEqualTo(ClassHash.TTL, Calendar.getInstance().getTimeInMillis())
                     .findFirst();
              ret = obj == null ? null : ClassHash.realmToObject(obj, mCollectionItemClass);
@@ -242,19 +246,17 @@ public class RealmCache<T extends GenericJson> implements ICache<T> {
     @Override
     public List<T> save(Iterable<T> items) {
         DynamicRealm mRealm = mCacheManager.getDynamicRealm();
-
         List<T> ret = new ArrayList<T>();
         try{
             mRealm.beginTransaction();
             for (T item : items){
-                item.put("_id", insertOrUpdate(item, mRealm));
+                item.put(ID, insertOrUpdate(item, mRealm));
                 ret.add(item);
             }
             mRealm.commitTransaction();
         } finally {
             mRealm.close();
         }
-        
         return ret;
     }
 
@@ -263,27 +265,37 @@ public class RealmCache<T extends GenericJson> implements ICache<T> {
     @Override
     public T save(T item) {
         DynamicRealm mRealm = mCacheManager.getDynamicRealm();
-
         try{
             mRealm.beginTransaction();
-            item.put("_id", insertOrUpdate(item, mRealm));
+            item.put(ID, insertOrUpdate(item, mRealm));
             mRealm.commitTransaction();
         } finally {
             mRealm.close();
         }
-       
         return item;
     }
 
     @Override
     public int delete(Query query) {
-        DynamicRealm mRealm = mCacheManager.getDynamicRealm();
-        int ret = 0;
+        DynamicRealm realm = mCacheManager.getDynamicRealm();
+        int i;
         try {
+            i = delete(realm, query, mCollection);
+            delete(realm, query, TableNameManager.getShortName(mCollection, realm) + ACL);
+            delete(realm, query, TableNameManager.getShortName(mCollection, realm) + KMD);
+        } finally {
+            realm.close();
+        }
+        return i;
+    }
+
+    private int delete(DynamicRealm realm, Query query, String tableName) {
+
+        int ret;
             if (!isQueryContainsInOperator(query.getQueryFilterMap())) {
-                mRealm.beginTransaction();
-                
-                RealmQuery<DynamicRealmObject> realmQuery = mRealm.where(TableNameManager.getShortName(mCollection, mRealm));
+                realm.beginTransaction();
+
+                RealmQuery<DynamicRealmObject> realmQuery = realm.where(TableNameManager.getShortName(tableName, realm));
                 QueryHelper.prepareRealmQuery(realmQuery, query.getQueryFilterMap());
 
                 RealmResults result = realmQuery.findAll();
@@ -292,19 +304,18 @@ public class RealmCache<T extends GenericJson> implements ICache<T> {
                 int limit = query.getLimit();
                 int skip = query.getSkip();
 
-                if (limit > 0)
-                {
+                if (limit > 0) {
                     // limit modifier has been applied, so take a subset of the Realm result set
                     int endIndex = Math.min(ret, (skip+limit));
                     List<DynamicRealmObject> subresult = result.subList(skip, endIndex);
                     List<String> ids = new ArrayList<String>();
                     ret = subresult.size();
                     for (DynamicRealmObject id : subresult) {
-                        ids.add((String)id.get("_id"));
+                        ids.add((String)id.get(ID));
                     }
-                    mRealm.commitTransaction();
+                    realm.commitTransaction();
                     if (ids.size() > 0) {
-                        this.delete(ids);
+                        this.delete(realm, ids, mCollection);
                     }
                 } else if (skip > 0) {
                     // only skip modifier has been applied, so take a subset of the Realm result set
@@ -313,18 +324,18 @@ public class RealmCache<T extends GenericJson> implements ICache<T> {
                         List<String> ids = new ArrayList<String>();
                         ret = subresult.size();
                         for (DynamicRealmObject id : subresult) {
-                            ids.add((String) id.get("_id"));
+                            ids.add((String) id.get(ID));
                         }
-                        mRealm.commitTransaction();
-                        this.delete(ids);
-                    }
-                    else {
+                        realm.commitTransaction();
+                        this.delete(realm, ids, mCollection);
+                    } else {
+                        realm.commitTransaction();
                         ret = 0;
                     }
                 } else {
                     // no skip or limit applied to query, so delete all results from Realm
                     result.deleteAllFromRealm();
-                    mRealm.commitTransaction();
+                    realm.commitTransaction();
                 }
 
             } else {
@@ -332,69 +343,75 @@ public class RealmCache<T extends GenericJson> implements ICache<T> {
                 ret = list.size();
                 List<String> ids = new ArrayList<>();
                 for (T id : list) {
-                    ids.add((String)id.get("_id"));
+                    ids.add((String)id.get(ID));
                 }
-                delete(ids);
+                delete(realm, ids, mCollection);
                 return ret;
             }
 
-        } finally {
-            mRealm.close();
-        }
         return ret;
     }
 
     @Override
     public int delete(Iterable<String> ids) {
-        DynamicRealm mRealm = mCacheManager.getDynamicRealm();
+        DynamicRealm realm = mCacheManager.getDynamicRealm();
+        int i;
+        try {
+             i = delete(realm, ids, mCollection);
+            delete(realm, ids, TableNameManager.getShortName(mCollection, realm) + ACL);
+            delete(realm, ids, TableNameManager.getShortName(mCollection, realm) + KMD);
+        } finally {
+            realm.close();
+        }
+        return i;
+    }
 
+    private int delete(DynamicRealm realm, Iterable<String> ids, String tableName) {
         int ret = 0;
-
-        try{
-            mRealm.beginTransaction();
-            RealmQuery<DynamicRealmObject> query = mRealm.where(TableNameManager.getShortName(mCollection, mRealm))
+        if (ids.iterator().hasNext()) {
+            realm.beginTransaction();
+            RealmQuery<DynamicRealmObject> query = realm.where(TableNameManager.getShortName(tableName, realm))
                     .greaterThanOrEqualTo(ClassHash.TTL, Long.MAX_VALUE)
                     .beginGroup();
             Iterator<String> iterator = ids.iterator();
-            if (iterator.hasNext()){
-                query.equalTo("_id", iterator.next());
-                for ( ; iterator.hasNext(); ){
-                    query.or().equalTo("_id", iterator.next());
+            if (iterator.hasNext()) {
+                query.equalTo(ID, iterator.next());
+                for (; iterator.hasNext(); ) {
+                    query.or().equalTo(ID, iterator.next());
                 }
-
             }
             query.endGroup();
-
             RealmResults result = query.findAll();
             ret = result.size();
             result.deleteAllFromRealm();
-            mRealm.commitTransaction();
-        } finally {
-            mRealm.close();
+            realm.commitTransaction();
         }
-
         return ret;
     }
 
     @Override
     public int delete(String id) {
-        DynamicRealm mRealm = mCacheManager.getDynamicRealm();
-
-
-        int ret = 0;
-
-        try{
-            mRealm.beginTransaction();
-            RealmQuery<DynamicRealmObject> query = mRealm.where(TableNameManager.getShortName(mCollection, mRealm))
-                    .equalTo("_id", id);
-            RealmResults realmResults = query.findAll();
-            ret = realmResults.size();
-            realmResults.deleteAllFromRealm();
-            mRealm.commitTransaction();
+        DynamicRealm realm = mCacheManager.getDynamicRealm();
+        int i;
+        try {
+            i = delete(realm, id, mCollection);
+            delete(realm, id, TableNameManager.getShortName(mCollection, realm) + ACL);
+            delete(realm, id, TableNameManager.getShortName(mCollection, realm) + KMD);
         } finally {
-            mRealm.close();
+            realm.close();
         }
-        
+        return i;
+    }
+
+    private int delete(DynamicRealm realm, String id, String tableName) {
+        int ret = 0;
+        realm.beginTransaction();
+        RealmQuery<DynamicRealmObject> query = realm.where(TableNameManager.getShortName(tableName, realm))
+                .equalTo(ID, id);
+        RealmResults realmResults = query.findAll();
+        ret = realmResults.size();
+        realmResults.deleteAllFromRealm();
+        realm.commitTransaction();
         return ret;
     }
 
@@ -404,7 +421,6 @@ public class RealmCache<T extends GenericJson> implements ICache<T> {
 
     public void clear(){
         DynamicRealm mRealm = mCacheManager.getDynamicRealm();
-
         try {
             mRealm.beginTransaction();
             mRealm.where(TableNameManager.getShortName(mCollection, mRealm))
@@ -508,9 +524,9 @@ public class RealmCache<T extends GenericJson> implements ICache<T> {
 
         ClassHash.saveData(mCollection, mRealm, mCollectionItemClass, clone);
 
-        item.set("_id", clone.get("_id"));
+        item.set(ID, clone.get(ID));
 
-        return item.get("_id").toString();
+        return item.get(ID).toString();
     }
 
 
@@ -604,7 +620,7 @@ public class RealmCache<T extends GenericJson> implements ICache<T> {
                             for (T item : newItems) {
                                 isItemExist = false;
                                 for(T oldItem : retCopy) {
-                                    if ((oldItem.get("_id")).equals(item.get("_id"))) {
+                                    if ((oldItem.get(ID)).equals(item.get(ID))) {
                                         isItemExist = true;
                                         break;
                                     }
